rapid form submission FIXED

+++ resources/views/dashboard/company.blade.php
@@ -241,12 +241,15 @@
       });
 
       //Dlt Modal Btn
+      let data = null
       $('#deletConfirmation .modal-footer').on('click' , '#btn-dlt' , function (e) {
+        if( data == $(this).attr('data-id')) return;
+        data = $(this).attr('data-id')
           axios.delete(`companys/${$(this).attr('data-id')}`)
           .then(function (response){
+              $('.btn-close-dlt').click();
               table.draw(false);
               displayToast(response , "success")
-              $('.btn-close-dlt').click();
           });
       });
